refactor(dev): move browser plugins into browserjs namespace
